requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
python-dotenv==1.0.0
geopy==2.3.0
requests==2.31.0
